Split one test into two
split test as what it was asking was ambiguous in regards to the readMe before. students were told they can either save the event in a variable or run the .persist() method. The test only checked for the .persist() method being used and was wrapped up with a test that checked whether the event was correctly passed to the callback. Since students could successfully complete the test by assigning to a variable instead of using .persist(), it was not clear to them why they were failing the test. 

Test now split into two, with the assertion of correct functionality of the callback separated from the assertion that the .persist() method is being invoked.